+++ test-schemas-domain-classes/src/main/scala/testdomains/hierarchical/edges/EdgeTypes.scala
@@ -6,4 +6,4 @@ object ConnectedTo {
 }
 
 class ConnectedTo(src_4762: flatgraph.GNode, dst_4762: flatgraph.GNode, subSeq_4862: Int, property_4862: Any)
-    extends flatgraph.Edge(src_4762, dst_4762, 0.toShort, subSeq_4862, property_4862) {}
+    extends flatgraph.Edge(src_4762, dst_4762, 0, subSeq_4862, property_4862) {}
